+++ backend/ee/danswer/danswerbot/slack/handlers/handle_standard_answers.py
@@ -21,11 +21,11 @@ from danswer.db.chat import get_or_create_root_message
 from danswer.db.models import Prompt
 from danswer.db.models import SlackBotConfig
 from danswer.db.models import StandardAnswer as StandardAnswerModel
-from danswer.server.manage.models import StandardAnswer as PydanticStandardAnswer
 from danswer.utils.logger import DanswerLoggingAdapter
 from danswer.utils.logger import setup_logger
 from ee.danswer.db.standard_answer import fetch_standard_answer_categories_by_names
 from ee.danswer.db.standard_answer import find_matching_standard_answers
+from ee.danswer.server.manage.models import StandardAnswer as PydanticStandardAnswer
 
 logger = setup_logger()
